DiffusionSolver moved into samurai
