+++ include/samurai/utils.hpp
@@ -131,4 +131,30 @@ namespace samurai
     {
         return static_cast<R>(static_cast<std::ptrdiff_t>(a) - static_cast<std::ptrdiff_t>(b));
     }
+
+    //------------------//
+    // Error management //
+    //------------------//
+    
+    void error(std::string msg)
+    {
+        std::string beginRed = "\033[1;31m";
+        std::string endColor = "\033[0m";
+        std::cout << beginRed << "Error: " << msg << endColor << std::endl;
+    }
+    void fatal_error(std::string msg)
+    {
+        std::string beginRed = "\033[1;31m";
+        std::string endColor = "\033[0m";
+        std::cout << beginRed << "Error: " << msg << endColor << std::endl;
+        std::cout << "------------------------- FAILURE -------------------------" << std::endl;
+        assert(false);
+        exit(EXIT_FAILURE);
+    }
+    void warning(std::string msg)
+    {
+        std::string beginYellow = "\033[1;33m";
+        std::string endColor = "\033[0m";
+        std::cout << beginYellow << "Warning: " << msg << endColor << std::endl;
+    }
 } // namespace samurai
